Add no data exception
